services/painter/requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
Flask==2.1.2
fonttools==4.33.3
gcloud==0.18.3
google-api-core==2.7.3
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.3
oauth2client==4.1.3
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
requests==2.27.1
rsa==4.8
scipy==1.8.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
uuid==1.30
Werkzeug==2.1.2
zipp==3.8.0
